+++ RELEASES.md
@@ -4,6 +4,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * prov-template-compiler: StatementCompilerAction2: refined past code generator for Attribution and Association
 * prov-template-compiler: javascript emitter updates to run workflow in js
 * prov-template-compiler: rust emitter updates to run workflow in rust
+* prov-template-compiler: CompilerBeanMerger, and PAST translation to Rust
 
 `2.2.3`
 * prov-template: update to instantiation